Pyatogui trading script for MacBook trading automation. buy and sel
The script has been tested on Mac functionality. Has not been updated with python for windows. The script goal is to switch from buy to sell While closing orders in succession. The script also listens to Web hooks on port 80 on telebit or any ssh and forwards to port 3000 for the script which is listening.

Buy side needs alerts from trading view to be  name “ buy”and alert “ buy” from tradingview webhooks for the script to understand. Also the placement of clicks is editable at the top of the screen. Happy trading.